Move common image validation functions into utils
