+++ weasyprint/css/validation/utils.py
@@ -14,7 +14,10 @@ import functools
 import math
 from urllib.parse import urljoin
 
+from tinycss2.color3 import parse_color
+
 from .. import computed_values
+from ...images import LinearGradient, RadialGradient
 from ...urls import iri_to_uri, url_is_absolute
 from ..properties import Dimension
 
@@ -39,9 +42,45 @@ RESOLUTION_TO_DPPX = {
 LENGTH_UNITS = (
     set(computed_values.LENGTHS_TO_PIXELS) | set(['ex', 'em', 'ch', 'rem']))
 
+# Constants about background positions
+ZERO_PERCENT = Dimension(0, '%')
+FIFTY_PERCENT = Dimension(50, '%')
+HUNDRED_PERCENT = Dimension(100, '%')
+BACKGROUND_POSITION_PERCENTAGES = {
+    'top': ZERO_PERCENT,
+    'left': ZERO_PERCENT,
+    'center': FIFTY_PERCENT,
+    'bottom': HUNDRED_PERCENT,
+    'right': HUNDRED_PERCENT,
+}
+
+# Direction keywords used for gradients
+DIRECTION_KEYWORDS = {
+    # ('angle', radians)  0 upwards, then clockwise
+    ('to', 'top'): ('angle', 0),
+    ('to', 'right'): ('angle', math.pi / 2),
+    ('to', 'bottom'): ('angle', math.pi),
+    ('to', 'left'): ('angle', math.pi * 3 / 2),
+    # ('corner', keyword)
+    ('to', 'top', 'left'): ('corner', 'top_left'),
+    ('to', 'left', 'top'): ('corner', 'top_left'),
+    ('to', 'top', 'right'): ('corner', 'top_right'),
+    ('to', 'right', 'top'): ('corner', 'top_right'),
+    ('to', 'bottom', 'left'): ('corner', 'bottom_left'),
+    ('to', 'left', 'bottom'): ('corner', 'bottom_left'),
+    ('to', 'bottom', 'right'): ('corner', 'bottom_right'),
+    ('to', 'right', 'bottom'): ('corner', 'bottom_right'),
+}
+
 
 class InvalidValues(ValueError):
     """Invalid or unsupported values for a known CSS property."""
+
+
+class CenterKeywordFakeToken(object):
+    type = 'ident'
+    lower_value = 'center'
+    unit = None
 
 
 def split_on_comma(tokens):
@@ -68,6 +107,16 @@ def remove_whitespace(tokens):
     return tuple(
         token for token in tokens
         if token.type not in ('whitespace', 'comment'))
+
+
+def safe_urljoin(base_url, url):
+    if url_is_absolute(url):
+        return iri_to_uri(url)
+    elif base_url:
+        return iri_to_uri(urljoin(base_url, url))
+    else:
+        raise InvalidValues(
+            'Relative URI reference without a base URI: %r' % url)
 
 
 def comma_separated_list(function):
@@ -129,7 +178,153 @@ def single_token(function):
     return single_token_validator
 
 
+def parse_linear_gradient_parameters(arguments):
+    first_arg = arguments[0]
+    if len(first_arg) == 1:
+        angle = get_angle(first_arg[0])
+        if angle is not None:
+            return ('angle', angle), arguments[1:]
+    else:
+        result = DIRECTION_KEYWORDS.get(tuple(map(get_keyword, first_arg)))
+        if result is not None:
+            return result, arguments[1:]
+    return ('angle', math.pi), arguments  # Default direction is 'to bottom'
+
+
+def parse_2d_position(tokens):
+    """Common syntax of background-position and transform-origin."""
+    if len(tokens) == 1:
+        tokens = [tokens[0], CenterKeywordFakeToken]
+    elif len(tokens) != 2:
+        return None
+
+    token_1, token_2 = tokens
+    length_1 = get_length(token_1, percentage=True)
+    length_2 = get_length(token_2, percentage=True)
+    if length_1 and length_2:
+        return length_1, length_2
+    keyword_1, keyword_2 = map(get_keyword, tokens)
+    if length_1 and keyword_2 in ('top', 'center', 'bottom'):
+        return length_1, BACKGROUND_POSITION_PERCENTAGES[keyword_2]
+    elif length_2 and keyword_1 in ('left', 'center', 'right'):
+            return BACKGROUND_POSITION_PERCENTAGES[keyword_1], length_2
+    elif (keyword_1 in ('left', 'center', 'right') and
+          keyword_2 in ('top', 'center', 'bottom')):
+        return (BACKGROUND_POSITION_PERCENTAGES[keyword_1],
+                BACKGROUND_POSITION_PERCENTAGES[keyword_2])
+    elif (keyword_1 in ('top', 'center', 'bottom') and
+          keyword_2 in ('left', 'center', 'right')):
+        # Swap tokens. They need to be in (horizontal, vertical) order.
+        return (BACKGROUND_POSITION_PERCENTAGES[keyword_2],
+                BACKGROUND_POSITION_PERCENTAGES[keyword_1])
+
+
+def parse_background_position(tokens):
+    """Parse background position.
+
+    See http://dev.w3.org/csswg/css3-background/#the-background-position
+
+    """
+    result = parse_2d_position(tokens)
+    if result is not None:
+        pos_x, pos_y = result
+        return 'left', pos_x, 'top', pos_y
+
+    if len(tokens) == 4:
+        keyword_1 = get_keyword(tokens[0])
+        keyword_2 = get_keyword(tokens[2])
+        length_1 = get_length(tokens[1], percentage=True)
+        length_2 = get_length(tokens[3], percentage=True)
+        if length_1 and length_2:
+            if (keyword_1 in ('left', 'right') and
+                    keyword_2 in ('top', 'bottom')):
+                return keyword_1, length_1, keyword_2, length_2
+            if (keyword_2 in ('left', 'right') and
+                    keyword_1 in ('top', 'bottom')):
+                return keyword_2, length_2, keyword_1, length_1
+
+    if len(tokens) == 3:
+        length = get_length(tokens[2], percentage=True)
+        if length is not None:
+            keyword = get_keyword(tokens[1])
+            other_keyword = get_keyword(tokens[0])
+        else:
+            length = get_length(tokens[1], percentage=True)
+            other_keyword = get_keyword(tokens[2])
+            keyword = get_keyword(tokens[0])
+
+        if length is not None:
+            if other_keyword == 'center':
+                if keyword in ('top', 'bottom'):
+                    return 'left', FIFTY_PERCENT, keyword, length
+                if keyword in ('left', 'right'):
+                    return keyword, length, 'top', FIFTY_PERCENT
+            elif (keyword in ('left', 'right') and
+                    other_keyword in ('top', 'bottom')):
+                return keyword, length, other_keyword, ZERO_PERCENT
+            elif (keyword in ('top', 'bottom') and
+                    other_keyword in ('left', 'right')):
+                return other_keyword, ZERO_PERCENT, keyword, length
+
+
+def parse_radial_gradient_parameters(arguments):
+    shape = None
+    position = None
+    size = None
+    size_shape = None
+    stack = arguments[0][::-1]
+    while stack:
+        token = stack.pop()
+        keyword = get_keyword(token)
+        if keyword == 'at':
+            position = parse_background_position(stack[::-1])
+            if position is None:
+                return
+            break
+        elif keyword in ('circle', 'ellipse') and shape is None:
+            shape = keyword
+        elif keyword in ('closest-corner', 'farthest-corner',
+                         'closest-side', 'farthest-side') and size is None:
+            size = 'keyword', keyword
+        else:
+            if stack and size is None:
+                length_1 = get_length(token, percentage=True)
+                length_2 = get_length(stack[-1], percentage=True)
+                if None not in (length_1, length_2):
+                    size = 'explicit', (length_1, length_2)
+                    size_shape = 'ellipse'
+                    stack.pop()
+            if size is None:
+                length_1 = get_length(token)
+                if length_1 is not None:
+                    size = 'explicit', (length_1, length_1)
+                    size_shape = 'circle'
+            if size is None:
+                return
+    if (shape, size_shape) in (('circle', 'ellipse'), ('circle', 'ellipse')):
+        return
+    return (
+        shape or size_shape or 'ellipse',
+        size or ('keyword', 'farthest-corner'),
+        position or ('left', FIFTY_PERCENT, 'top', FIFTY_PERCENT),
+        arguments[1:])
+
+
+def parse_color_stop(tokens):
+    if len(tokens) == 1:
+        color = parse_color(tokens[0])
+        if color is not None:
+            return color, None
+    elif len(tokens) == 2:
+        color = parse_color(tokens[0])
+        position = get_length(tokens[1], negative=True, percentage=True)
+        if color is not None and position is not None:
+            return color, position
+    raise InvalidValues
+
+
 def get_length(token, negative=True, percentage=False):
+    """Parse a <length> token."""
     if percentage and token.type == 'percentage':
         if negative or token.value >= 0:
             return Dimension(token.value, '%')
@@ -141,7 +336,7 @@ def get_length(token, negative=True, percentage=False):
 
 
 def get_angle(token):
-    """Return the value in radians of an <angle> token, or None."""
+    """Parse an <angle> token in radians."""
     if token.type == 'dimension':
         factor = ANGLE_TO_RADIANS.get(token.unit)
         if factor is not None:
@@ -149,18 +344,39 @@ def get_angle(token):
 
 
 def get_resolution(token):
-    """Return the value in dppx of a <resolution> token, or None."""
+    """Parse a <resolution> token in ddpx."""
     if token.type == 'dimension':
         factor = RESOLUTION_TO_DPPX.get(token.unit)
         if factor is not None:
             return token.value * factor
 
 
-def safe_urljoin(base_url, url):
-    if url_is_absolute(url):
-        return iri_to_uri(url)
-    elif base_url:
-        return iri_to_uri(urljoin(base_url, url))
-    else:
-        raise InvalidValues(
-            'Relative URI reference without a base URI: %r' % url)
+def get_image(token, base_url):
+    """Parse an <image> token."""
+    if token.type != 'function':
+        if get_keyword(token) == 'none':
+            return 'none', None
+        if token.type == 'url':
+            return 'url', safe_urljoin(base_url, token.value)
+        return
+    arguments = split_on_comma(remove_whitespace(token.arguments))
+    name = token.lower_name
+    if name in ('linear-gradient', 'repeating-linear-gradient'):
+        direction, color_stops = parse_linear_gradient_parameters(arguments)
+        if color_stops:
+            return 'linear-gradient', LinearGradient(
+                [parse_color_stop(stop) for stop in color_stops],
+                direction, 'repeating' in name)
+    elif name in ('radial-gradient', 'repeating-radial-gradient'):
+        result = parse_radial_gradient_parameters(arguments)
+        if result is not None:
+            shape, size, position, color_stops = result
+        else:
+            shape = 'ellipse'
+            size = 'keyword', 'farthest-corner'
+            position = 'left', FIFTY_PERCENT, 'top', FIFTY_PERCENT
+            color_stops = arguments
+        if color_stops:
+            return 'radial-gradient', RadialGradient(
+                [parse_color_stop(stop) for stop in color_stops],
+                shape, size, position, 'repeating' in name)
